requirements.txt
# Core Dependencies
numpy>=1.20.0,<2.0.0
pandas>=1.3.0,<2.0.0
scikit-learn>=1.0.0,<2.0.0
scipy>=1.7.0,<2.0.0
joblib>=1.1.0,<2.0.0
tqdm>=4.62.0,<5.0.0

# Optimization
hyperopt>=0.2.7,<0.3.0
optuna>=2.10.0,<3.0.0
bayesian-optimization>=1.2.0,<2.0.0

# Deep Learning (Optional)
torch>=1.10.0,<2.0.0
tensorflow>=2.8.0,<3.0.0
keras>=2.8.0,<3.0.0

# Feature Processing
category_encoders>=2.3.0,<3.0.0
feature-engine>=1.2.0,<2.0.0
nltk>=3.6.0,<4.0.0
spacy>=3.2.0,<4.0.0

# Model Interpretability
shap>=0.40.0,<0.41.0
lime>=0.2.0.1,<0.3.0
eli5>=0.11.0,<0.12.0

# Visualization
matplotlib>=3.4.0,<4.0.0
seaborn>=0.11.0,<0.12.0
plotly>=5.5.0,<6.0.0

# Metrics and Evaluation
scikit-learn>=1.0.0,<2.0.0  # Already included above, but ensures metric functionality
pandas>=1.3.0,<2.0.0  # Already included above
pandera>=0.8.0,<0.9.0
scikit-optimize>=0.9.0,<1.0.0  # Additional metrics and optimization utilities

# Utilities
pyyaml>=6.0,<7.0.0
python-dotenv>=0.19.0,<0.20.0
loguru>=0.5.3,<0.6.0

# Testing
pytest>=7.0.0,<8.0.0
pytest-cov>=3.0.0,<4.0.0
hypothesis>=6.0.0,<7.0.0

# Documentation
sphinx>=4.4.0,<5.0.0
sphinx-rtd-theme>=1.0.0,<2.0.0
nbsphinx>=0.8.8,<0.9.0
jupyter>=1.0.0,<2.0.0

# Type Checking
mypy>=0.930,<1.0.0
types-PyYAML>=6.0.0,<7.0.0
types-setuptools>=57.0.0,<58.0.0

# Code Quality
black>=22.0.0,<23.0.0
flake8>=4.0.0,<5.0.0
isort>=5.10.0,<6.0.0
pylint>=2.12.0,<3.0.0